Modify alert
