+++ ha/setup/create_pacemaker_resources.py
@@ -15,7 +15,6 @@
 # about this software or licensing, please email opensource@seagate.com or
 # cortx-questions@seagate.com.
 
-import os
 import argparse
 from ha.execute import SimpleCommand
 from cortx.utils.log import Log
@@ -332,63 +331,3 @@ def create_all_resources(cib_xml=const.CIB_FILE, push=True, **kwargs):
         Log.info("HA Rules: Successfully configured all HA resources and its configuration.")
     except Exception:
         raise CreateResourceError("Failed to populate cluster with resources")
-
-class ConfigAlertResource:
-
-    ALERT_SCRIPT_PATH = ["/usr/local/bin/pcmk_alert", "/usr/bin/pcmk_alert"]
-    ALERT_ID = "iem_alert"
-    RECIPIENT_KEY = "sender_type"
-    RECIPIENT_VALUE = "syslog"
-
-    def __init__(self):
-        self._process = SimpleCommand()
-
-    def is_alert_exists(self) -> bool:
-        """
-        Check if alert already exists.
-        """
-        Log.info("Checking pacemaker alert if already exists ...")
-        output, error, rc = self._process.run_cmd(f"pcs alert")
-        for line in output.split("\n"):
-            if ConfigAlertResource.ALERT_ID in line:
-                return True
-        return False
-
-    def create_alert(self):
-        """
-        Manage alert resource.
-        """
-        Log.info("Creating pacemaker alert ...")
-        if self.is_alert_exists():
-            self.delete_alert()
-        path: str = ConfigAlertResource._get_script()
-        self._process.run_cmd(f"pcs alert create id={ConfigAlertResource.ALERT_ID} description=send_iem_alerts path={path}")
-        self._process.run_cmd(f"pcs alert recipient add {ConfigAlertResource.ALERT_ID} id={ConfigAlertResource.RECIPIENT_KEY} value={ConfigAlertResource.RECIPIENT_VALUE}")
-        Log.info(f"Alert {ConfigAlertResource.ALERT_ID} created successfully.")
-
-    def delete_alert(self):
-        """
-        Delete alert on current node.
-        """
-        if not self.is_alert_exists():
-            return
-        Log.info("Deleating pacemaker alert ...")
-        self._process.run_cmd(f"pcs alert remove {ConfigAlertResource.ALERT_ID}")
-        Log.info(f"Alert {ConfigAlertResource.ALERT_ID} is deleted")
-
-    @staticmethod
-    def _get_script() -> str:
-        """
-        Get alert script
-
-        Raises:
-            CreateResourceConfigError: Raise error if script missing
-        """
-        path: str = None
-        for script in ConfigAlertResource.ALERT_SCRIPT_PATH:
-            if os.path.isfile(script):
-                path = script
-                break
-        if path is None:
-            raise CreateResourceConfigError(f"Failed to create alert missing {str(ConfigAlertResource.ALERT_SCRIPT_PATH)} file.")
-        return path
