+++ ha/setup/ha_setup.py
@@ -32,7 +32,7 @@ from ha.execute import SimpleCommand
 from ha import const
 from ha.const import STATUSES
 from ha.setup.create_pacemaker_resources import create_all_resources
-from ha.setup.create_pacemaker_resources import ConfigAlertResource
+from ha.setup.pcs_config.alert_config import AlertConfig
 from ha.setup.post_disruptive_upgrade import perform_post_upgrade
 from ha.core.cluster.cluster_manager import CortxClusterManager
 from ha.core.config.config_manager import ConfigManager
@@ -311,7 +311,7 @@ class ConfigCmd(Cmd):
         Init method.
         """
         super().__init__(args)
-        self._alert_resource = ConfigAlertResource()
+        self._alert_config = AlertConfig()
 
     def process(self):
         """
@@ -360,11 +360,11 @@ class ConfigCmd(Cmd):
                 try:
                     self._create_cluster(cluster_name, cluster_user, cluster_secret, node_name)
                     self._create_resource(s3_instances=s3_instances, mgmt_info=mgmt_info)
-                    self._alert_resource.create_alert()
+                    self._alert_config.create_alert()
                     self._confstore.set(f"{const.CLUSTER_CONFSTORE_NODES_KEY}/{node_name}")
                 except Exception as e:
                     Log.error(f"Cluster creation failed; destroying the cluster. Error: {e}")
-                    self._alert_resource.delete_alert()
+                    self._alert_config.delete_alert()
                     output = self._execute.run_cmd(const.PCS_CLUSTER_DESTROY)
                     Log.error(f"Cluster destroyed. Output: {output}")
                     # Delete the node from nodelist if it was added in the store
